Condense the Yes/No dialog box to a single component
